# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مرشد الكتب الذكي - AI weather-based book recommender
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/book-preview:
    post:
      operationId: createAiBookPreview
      tags: [ai]
      summary: Preview books from a natural-language request
      description: Interprets a reader's free-form request with the current weather and mood context, then returns books for review before applying them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiBookPreviewInput"
      responses:
        "200":
          description: Preview recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiBookPreview"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/transcribe:
    post:
      operationId: transcribeVoice
      tags: [ai]
      summary: Transcribe a voice reading request
      description: Converts a short browser recording into editable text. Audio is processed in memory and is not stored.
      requestBody:
        required: true
        content:
          audio/webm:
            schema:
              type: string
              format: binary
          audio/wav:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Transcribed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transcription"
        "400":
          description: Missing or unsupported audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
            default: /
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL with a post-logout redirect back to the validated `returnTo` path.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AiBookPreviewInput:
      type: object
      properties:
        prompt:
          type: string
          minLength: 3
          maxLength: 1200
        mode:
          type: string
          enum: [custom, hidden_gem, trending, classic, excerpt]
          default: custom
        searchRound:
          type: integer
          minimum: 1
          maximum: 50
          default: 1
        city:
          type: string
        mood:
          type: string
        category:
          type: string
        weatherCondition:
          type: string
        temperature:
          type: number
        timeOfDay:
          type: string
        cityBookLang:
          type: string
          enum: [ar, en, fr]
        language:
          type: string
          enum: [ar, en, fr]
        excludeBooks:
          type: array
          items:
            type: string
        bookLength:
          type: string
        ageGroup:
          type: string
        season:
          type: string
      required:
        - prompt
    AiBookPreviewBook:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        category:
          type: string
        subcategory:
          type: string
        rating:
          type: string
        reason:
          type: string
        quote:
          type: string
        atmosphere:
          type: string
        length:
          type: string
        publishedYear:
          type: integer
          minimum: 0
        publishedMonth:
          type: integer
          minimum: 1
          maximum: 12
        tags:
          type: array
          items:
            type: string
        fitScore:
          type: integer
          minimum: 0
          maximum: 100
      required:
        - title
        - author
        - category
        - subcategory
        - rating
        - reason
        - quote
        - tags
        - fitScore
    AiBookPreview:
      type: object
      properties:
        interpretedRequest:
          type: string
        summary:
          type: string
        books:
          type: array
          minItems: 1
          maxItems: 5
          items:
            $ref: "#/components/schemas/AiBookPreviewBook"
      required:
        - interpretedRequest
        - summary
        - books
    Transcription:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string