# lib/api-zod/src/generated/types/aiBookPreviewBook.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for مرشد الكتب الذكي - AI weather-based book recommender
 * OpenAPI spec version: 0.1.0
 */

export interface AiBookPreviewBook {
  title: string;
  author: string;
  category: string;
  subcategory: string;
  rating: string;
  reason: string;
  quote: string;
  atmosphere?: string;
  length?: string;
  /** @minimum 0 */
  publishedYear?: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  publishedMonth?: number;
  tags: string[];
  /**
     * @minimum 0
     * @maximum 100
     */
  fitScore: number;
}