# lib/api-zod/src/generated/types/aiBookPreviewInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for مرشد الكتب الذكي - AI weather-based book recommender
 * OpenAPI spec version: 0.1.0
 */
import type { AiBookPreviewInputCityBookLang } from './aiBookPreviewInputCityBookLang';
import type { AiBookPreviewInputLanguage } from './aiBookPreviewInputLanguage';
import type { AiBookPreviewInputMode } from './aiBookPreviewInputMode';

export interface AiBookPreviewInput {
  /**
     * @minLength 3
     * @maxLength 1200
     */
  prompt: string;
  mode?: AiBookPreviewInputMode;
  /**
     * @minimum 1
     * @maximum 50
     */
  searchRound?: number;
  city?: string;
  mood?: string;
  category?: string;
  weatherCondition?: string;
  temperature?: number;
  timeOfDay?: string;
  cityBookLang?: AiBookPreviewInputCityBookLang;
  language?: AiBookPreviewInputLanguage;
  excludeBooks?: string[];
  bookLength?: string;
  ageGroup?: string;
  season?: string;
}