# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for مرشد الكتب الذكي - AI weather-based book recommender
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}